refactor: Move `new-readonly`components
