+++ frontend/src/app/sessions/user-sessions-wrapper/create-sessions/create-readonly-session/new-readonly-session-dialog.component.ts
@@ -15,8 +15,8 @@ import {
   ModelService,
 } from 'src/app/projects/models/service/model.service';
 import { GetGitModel } from 'src/app/projects/project-detail/model-overview/model-detail/git-model.service';
-import { ModelOptions } from 'src/app/sessions/new-readonly-session-dialog/new-readonly-model-options/new-readonly-model-options.component';
-import { SessionService } from '../service/session.service';
+import { SessionService } from 'src/app/sessions/service/session.service';
+import { ModelOptions } from 'src/app/sessions/user-sessions-wrapper/create-sessions/create-readonly-session/new-readonly-model-options/new-readonly-model-options.component';
 
 @UntilDestroy()
 @Component({
